Last push
float->double per trollare
messo qualche inline (funzioni piccole)

+++ scene.h
@@ -14,30 +14,30 @@ using Triangle = std::array<size_t, 3>;
 class Vertex{
     private:
         // Constant, not-perspective projected coordinates, used because different pipelines (and different rendering behaviours) may re-use vertices
-        const float x_, y_, z_ = 0.0f;
+        const double x_, y_, z_ = 0.0f;
         // Perspective projected coordinates, re-calculated every time a pipeline renders, using the Vertex
-        float ndx_, ndy_, ndz_ = 0.0f;
+        double ndx_, ndy_, ndz_ = 0.0f;
 
     public:
         // Constructors
         Vertex() = default;
-        Vertex(const float x, const float y, const float z): x_(x), y_(y), z_(z){}
+        Vertex(const double x, const double y, const double z): x_(x), y_(y), z_(z){}
         Vertex(const Vertex& v) : x_(v.x_), y_(v.y_), z_(v.z_), ndx_(v.ndx_), ndy_(v.ndy_), ndz_(v.ndz_){}
 
         // Getters of the constant coordinates
-        float getX()  {return x_;}
-        float getY()  {return y_;}
-        float getZ()  {return z_;}
+        double getX()  {return x_;}
+        double getY()  {return y_;}
+        double getZ()  {return z_;}
 
         // Getters of the pipeline-computed coordinates
-        float getNdx()  {return ndx_;}
-        float getNdy()  {return ndy_;}
-        float getNdz()  {return ndz_;}
+        double getNdx()  {return ndx_;}
+        double getNdy()  {return ndy_;}
+        double getNdz()  {return ndz_;}
 
         // Setters
-        void setNdx(const float x){ndx_ = x;}
-        void setNdy(const float y){ndy_= y;}
-        void setNdz(const float z){ndz_ = z;}
+        void setNdx(const double x){ndx_ = x;}
+        void setNdy(const double y){ndy_= y;}
+        void setNdz(const double z){ndz_ = z;}
 
 
 };
